refactor: rename typo in category file
