+++ src/controllers/organisation/index.ts
@@ -1,4 +1,4 @@
-export * from './catergory'
+export * from './category'
 export * from './list'
 export * from './delete'
 export * from './update'
